fix: 컴파일 에러 수정

+++ src/main/java/com/sparta/foodorder/domain/store/domain/Store.java
@@ -145,7 +145,7 @@ public class Store extends BaseEntity {
         this.isActive = false;
     }
 
-    public void updateRating(Long ratingCount, Float ratingAvg) {
+    public void updateRating(Long ratingCount, Double ratingAvg) {
         if(ratingCount != null) this.ratingCount = ratingCount;
         if(ratingAvg != null) this.ratingAvg = ratingAvg;
     }
